Remove ScrollShim mixin
